Extract reusable module handling path history for back button support
